code cleanup and clippy refactor
