+++ hvat/src/hyperview/auth.rs
@@ -4,7 +4,7 @@ use oauth2::{
     TokenResponse, TokenUrl,
 };
 
-use crate::hyperview::cli_data::AppConfig;
+use super::cli_data::AppConfig;
 
 pub async fn get_auth_header_async(config: &AppConfig) -> Result<String> {
     // create client
